Pseudocode #56
