+++ src/main/java/nl/windesheim/ictm2o/peach/design/DPToevCompon.java
@@ -164,4 +164,6 @@ public class DPToevCompon extends JPanel implements ActionListener {
         }
 
     }
+
+
 }
